rpi/requirements.txt
ultralytics
numpy
opencv-python
paho-mqtt